Add files via upload
The dining problem states that five philosophers share a table they have to eat and think. In the table a bowl of rice and five chopsticks. Every philosopher requires a left and a right chopstick. A hungry philosopher can eat when the left and the right chopsticks are free, or a philosopher have to put their chopstick down and thinks .

This problem was designed to illustrate the challenges of avoiding deadlock, a system state in which no progress is possible.